refactor: use Upload namespace
