+++ packages/upload-client/src/constants.js
@@ -1,0 +1,1 @@
+export const REQUEST_RETRIES = 3
