+++ packages/upload-client/src/storage.js
@@ -1,79 +1,9 @@
-import { isDelegation } from '@ucanto/core'
-import { connect } from '@ucanto/client'
-import { CAR, CBOR, HTTP } from '@ucanto/transport'
-import * as DID from '@ipld/dag-ucan/did'
+import { CAR } from '@ucanto/transport'
 import { add as storeAdd } from '@web3-storage/access/capabilities/store'
-import { add as uploadAdd } from '@web3-storage/access/capabilities/upload'
 import retry, { AbortError } from 'p-retry'
-
-// Production
-const serviceURL = new URL(
-  'https://8609r1772a.execute-api.us-east-1.amazonaws.com'
-)
-const serviceDID = DID.parse(
-  'did:key:z6MkrZ1r5XBFZjBU34qyD8fueMbMRkKw17BZaq2ivKFjnz2z'
-)
-
-const RETRIES = 3
-
-const connection = connect({
-  id: serviceDID,
-  encoder: CAR,
-  decoder: CBOR,
-  channel: HTTP.open({
-    url: serviceURL,
-    method: 'POST',
-  }),
-})
-
-/**
- * Register an "upload" with the service. The issuer needs the `upload/add`
- * delegated capability.
- *
- * Required delegated capability proofs: `upload/add`
- *
- * @param {import('./types').InvocationConfig} invocationConfig Configuration
- * for the UCAN invocation. An object with `issuer` and `proofs`.
- *
- * The `issuer` is the signing authority that is issuing the UCAN
- * invocation(s). It is typically the user _agent_.
- *
- * The `proofs` are a set of capability delegations that prove the issuer
- * has the capability to perform the action.
- *
- * The issuer needs the `upload/add` delegated capability.
- * @param {import('multiformats/link').UnknownLink} root Root data CID for the DAG that was stored.
- * @param {import('./types').CARLink[]} shards CIDs of CAR files that contain the DAG.
- * @param {import('./types').RequestOptions} [options]
- */
-export async function registerUpload(
-  { issuer, proofs },
-  root,
-  shards,
-  options = {}
-) {
-  const capability = findCapability(proofs, serviceDID.did(), uploadAdd.can)
-  /** @type {import('@ucanto/interface').ConnectionView<import('./types').Service>} */
-  const conn = options.connection ?? connection
-  await retry(
-    async () => {
-      const result = await uploadAdd
-        .invoke({
-          issuer,
-          audience: serviceDID,
-          // @ts-expect-error expects did:${string} but cap with is ${string}:${string}
-          with: capability.with,
-          nb: {
-            root,
-            shards,
-          },
-        })
-        .execute(conn)
-      if (result?.error === true) throw result
-    },
-    { onFailedAttempt: console.warn, retries: options.retries ?? RETRIES }
-  )
-}
+import { serviceDID, connection } from './service.js'
+import { findCapability } from './utils.js'
+import { REQUEST_RETRIES } from './constants.js'
 
 /**
  * Store a DAG encoded as a CAR file. The issuer needs the `store/add`
@@ -116,7 +46,10 @@ export async function store({ issuer, proofs }, car, options = {}) {
         .execute(conn)
       return res
     },
-    { onFailedAttempt: console.warn, retries: options.retries ?? RETRIES }
+    {
+      onFailedAttempt: console.warn,
+      retries: options.retries ?? REQUEST_RETRIES,
+    }
   )
 
   if (result.error) {
@@ -149,7 +82,10 @@ export async function store({ issuer, proofs }, car, options = {}) {
         throw err
       }
     },
-    { onFailedAttempt: console.warn, retries: options.retries ?? RETRIES }
+    {
+      onFailedAttempt: console.warn,
+      retries: options.retries ?? REQUEST_RETRIES,
+    }
   )
 
   if (!res.ok) {
@@ -157,39 +93,4 @@ export async function store({ issuer, proofs }, car, options = {}) {
   }
 
   return link
-}
-
-/**
- * @param {import('@ucanto/interface').Proof[]} proofs
- * @param {import('@ucanto/interface').DID} audience
- * @param {import('@ucanto/interface').Ability} ability
- */
-function findCapability(proofs, audience, ability) {
-  let capability
-  for (const proof of proofs) {
-    if (!isDelegation(proof)) continue
-    if (proof.audience.did() !== audience) continue
-    capability = proof.capabilities.find((c) =>
-      capabilityMatches(c.can, ability)
-    )
-    if (capability) break
-  }
-  if (!capability) {
-    throw new Error(
-      `Missing proof of delegated capability "${
-        uploadAdd.can
-      }" for audience "${serviceDID.did()}"`
-    )
-  }
-  return capability
-}
-
-/**
- * @param {string} can
- * @param {import('@ucanto/interface').Ability} ability
- */
-function capabilityMatches(can, ability) {
-  return can === ability
-    ? true
-    : can.endsWith('*') && ability.startsWith(can.split('*')[0])
 }
